Rename for now Weather as WeatherMultiple, add Weather and Pomodoro component stub
