+++ src/components/App.js
@@ -8,6 +8,8 @@ import GoogleCalendar from './GoogleCalendar/GoogleCalendar';
 import Search from './Search/Search';
 import CurrencyConverter from './CurrencyConverter/Currency';
 import Weather from './Weather/Weather';
+import WeatherMultiple from './Weather/WeatherMultiple';
+import Pomodoro from './Pomodoro/Pomodoro';
 import './card.css';
 import banner from '../app_banner.jpeg';
 
@@ -19,13 +21,15 @@ function App() {
   return (
     <div className="App">
       <img src={banner} alt="banner" className="App-banner" />
+      <Weather className="App-Weather" />
       <Search className="App-Search" />
       <div className="App-features-grid">
         <CurrencyConverter API_KEY={CurrencyKey} />
-        <Weather IP_KEY={IpKey} API_KEY={WeatherKey} />
+        <Pomodoro />
         <GoogleCalendar />
       </div>
 
+      <WeatherMultiple IP_KEY={IpKey} API_KEY={WeatherKey} />
       <div className="App-react-context">
         <TestState>
           <StateReceiver />
