+++ src/components/Card/Card.css
@@ -5,7 +5,7 @@
   box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
   border-radius: 15px;
   padding: 20px;
-  min-height: 400px;
+  min-height: 200px;
   box-sizing: border-box;
   text-align: center;
 }
