+++ src/components/Card/Card.js
@@ -3,12 +3,17 @@ import { PropTypes } from 'prop-types';
 import './Card.css';
 
 function Card(props) {
-  const { children } = props;
-  return <div className="Card">{children}</div>;
+  const { children, className } = props;
+  return <div className={`Card ${className}`}>{children}</div>;
 }
 
 Card.propTypes = {
-  children: PropTypes.node.isRequired
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string
+};
+
+Card.defaultProps = {
+  className: ''
 };
 
 export default Card;
