+++ src/components/Pomodoro/Pomodoro.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Card from '../Card/Card';
+
+function Pomodoro() {
+  return <Card />;
+}
+
+export default Pomodoro;
